Move sail-derived code to a separate folder
This will make it easier to disable clang-format for these files and makes
it obvious that these files should not be edited since they are either
vendored from sail or generated.
