+++ test/regen_sail_c_code.sh
@@ -22,7 +22,7 @@ fi
 sail_pre_srcs="sail-riscv/model/prelude.sail sail-riscv/model/prelude_mapping.sail"
 sail128_srcs="$sail_pre_srcs sail-riscv/model/riscv_xlen64.sail sail-riscv/model/riscv_flen_D.sail src/cheri_prelude.sail src/cheri_types.sail src/cheri_prelude_128.sail src/cheri_cap_common.sail"
 sail64_srcs="$sail_pre_srcs sail-riscv/model/riscv_xlen32.sail sail-riscv/model/riscv_flen_F.sail src/cheri_prelude.sail src/cheri_types.sail src/cheri_prelude_64.sail src/cheri_cap_common.sail"
-output_dir=${SCRIPT_DIR}
+output_dir=${SCRIPT_DIR}/contrib
 keep_required_functions=(-c_preserve capToBits
 	-c_preserve capToMemBits
 	-c_preserve capBitsToCapability
@@ -38,8 +38,8 @@ keep_required_functions=(-c_preserve capToBits
 
 cd "$SAIL_RISCV_DIR"
 # -c_no_rts
-sail -c -c_no_main -c_prefix sailgen_ -c_specialize "${keep_required_functions[@]}" -verbose=2 -o "$output_dir/sail_compression_128" $sail128_srcs "$output_dir/compression_test.sail" -static
-sail -c -c_no_main -c_prefix sailgen_ -c_specialize "${keep_required_functions[@]}" -verbose=2 -o "$output_dir/sail_compression_64" $sail64_srcs "$output_dir/compression_test.sail" -static
+sail -c -c_no_main -c_prefix sailgen_ -c_specialize "${keep_required_functions[@]}" -verbose=2 -o "$output_dir/sail_compression_128" $sail128_srcs "${SCRIPT_DIR}/compression_test.sail" -static
+sail -c -c_no_main -c_prefix sailgen_ -c_specialize "${keep_required_functions[@]}" -verbose=2 -o "$output_dir/sail_compression_64" $sail64_srcs "${SCRIPT_DIR}/compression_test.sail" -static
 cd "$output_dir"
 
 for i in sail.h sail.c sail_failure.c sail_failure.h; do
